+ Updated DB to v21 + Added validateEmail command + Added server-dist/docs/phpBB Synchronization.pdf with instructions     for synchronizing user accounts between phpBB 2.0.22 and MW + Defection now requires a validated forum account if the server is     synching with phpBB + Added server-dist/lib/smtp.jar + Added integrated javamail to server + Completed phpBB account activation code + Unenrolling will no longer remove phpBB accounts (deletion will)

+++ trunk/src/server/mwmysql/PlayerHandler.java
@@ -151,6 +151,10 @@ public class PlayerHandler {
 	}
 	
 	public void deletePlayer(SPlayer p) {
+		deletePlayer(p, true);
+	}
+	
+	public void deletePlayer(SPlayer p, boolean deleteForumAccount) {
 		Statement stmt;
 		try {
 			// Remove armies
@@ -173,7 +177,7 @@ public class PlayerHandler {
 			rs.close();
 			
 			// Remove from phpBB database
-			if(CampaignMain.cm.isSynchingBB()) 
+			if(CampaignMain.cm.isSynchingBB() && deleteForumAccount) 
 				CampaignMain.cm.MySQL.deleteForumAccount(p.getForumID());
 			// This has its own set of issues right now - I need to talk to Orca before I put this in.
